Server refresh optimizations
Utilize new function 'RefreshServerList()' which obtains the list from the server.
Since the constructor of the server browser in code already does this, we don't need to 'refresh' when the VM is initialized as the VM uses the same globals shared with the browser in a thread safe manner.

GetServerCount() no longer does the http post operation, but instead obtains the size of the list vector directly in a thread safe manner.