Improve disconnect code
Removed global 'g_isAtMainMenu', and moved disconnect logic into the server creation thread. Also added an additional second of wait time to make the transition smoother, instead of instantly switching to the lobby as it couldn't even finish the sounds. So far this change appears to be robust, it always works (tested on dedicated and listen servers). The previous implementation tends to fail every now and then.